apis

+++ features/transaction/atc_common/src/main/java/com/tokopedia/atc_common/data/model/request/AddToCartOccRequestParams.kt
@@ -1,9 +1,7 @@
 package com.tokopedia.atc_common.data.model.request
 
-import com.google.gson.annotations.SerializedName
-
 data class AddToCartOccRequestParams(
-        var productId: Long,
+        var productId: Int,
         var shopId: Int,
         var quantity: Int,
         var warehouseId: Int = 0,
